+++ packages/flutter_translate/lib/translate_preferences.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/cupertino.dart';
+import 'dart:ui';
 
 abstract class ITranslatePreferences
 {